Organized files between html and other types
